pyproject.toml
[project]
name = "micro-paper"
version = "0.1.0"
description = "MLX Arena — micro & macro model evaluation"
requires-python = ">=3.12"
dependencies = [
    "mlx>=0.22",
    "mlx-lm>=0.21",
    "datasets",
    "typer",
    "python-dotenv",
    "openai",
]

[project.scripts]
micro-paper = "run_macro:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["macro", "micro", "tribe"]
